Update .replit to reenable the package manager button
If there is no previous configuration made, the package manager is hidden, and so, a user cannot install any libraries via pip.
I have included the contents of the default Python Template file because it works fine.
This issue is present in the other repos as well, it would be great if you could fix it in order to help devs. I, for one, downloaded the repo and solved locally, only to then copy and paste it to repl.it. I found out just today about this solution so I thought I'd share it!